+++ src/main/java/com/kimy1212/progressmeter/controller/dto/TodoTabResponse.java
@@ -2,7 +2,7 @@ package com.kimy1212.progressmeter.controller.dto;
 
 public record TodoTabResponse(
 
-		int todoTabId,
+		long todoTabId,
 
 		String todoTabName
