+++ src/main/java/com/kimy1212/progressmeter/infrastructure/repository/row/TodoTabRow.java
@@ -2,7 +2,7 @@ package com.kimy1212.progressmeter.infrastructure.repository.row;
 
 public record TodoTabRow(
 
-		int todoTabId,
+		long todoTabId,
 
 		String todoTabName
